cyl-task2-dev reformat code structure

+++ task2/client_script/upload_startup.py
@@ -9,10 +9,8 @@ context = zmq.Context()
 
 # create a ZMQ PUB socket
 socket = context.socket(zmq.PUB)
-socket.connect(f"tcp://{ZMQ_HOST}:{ZMQ_PORT}")
+socket.connect(f"tcp://{SERVER_HOST}:{ZMQ_PORT}")
 
-# set the topic for this publisher
-topic = ZMQ_TOPICS[1]
 
 def conn_check() -> bool:
     # Run the command to get the SSID of the currently connected network
@@ -29,26 +27,31 @@ def conn_check() -> bool:
         print(f"Currently connected to WIFI with ssid: {ssid}")
         return ssid == wifi_name
 
+
+# set the topic for this publisher
+topic = ZMQ_TOPIC
+
 while True:
     if not conn_check():
         time.sleep(RETRY_INTERVAL_SEC)
         continue
 
     # scan the folder and upload all files with the given topic
-    for filename in os.listdir(folder_path[1]):
-        filepath = os.path.join(folder_path[1], filename)
-        with open(filepath, "rb") as f:
-            data = f.read()
+    for filename in os.listdir(folder_path):
+        if filename.endswith(file_ends):
+            filepath = os.path.join(folder_path, filename)
+            with open(filepath, "rb") as f:
+                data = f.read()
 
-        try:
-            # send the data to the subscriber with the given topic
-            socket.send_multipart([topic.encode(), filename.encode(), data])
-            print(f"Published {filename} on topic {topic}")
-        except zmq.ZMQError as e:
-            print(f"Error: {e}, retrying in 5 seconds")
-            time.sleep(RETRY_INTERVAL_SEC)
-            break
+            try:
+                # send the data to the subscriber with the given topic
+                socket.send_multipart([topic.encode(), filename.encode(), data])
+                print(f"Published {filename} on topic {topic}")
+            except zmq.ZMQError as e:
+                print(f"Error: {e}, retrying in 5 seconds")
+                time.sleep(RETRY_INTERVAL_SEC)
+                break
 
-        os.remove(filepath)
+            os.remove(filepath)
 
     time.sleep(SCAN_INTERVAL_SEC)
